Fix POST handling.
In a bug caught by @henryjyc, which I introduced after misreading the
documentation, POST handlers expected their parameters either in the URL or as
'form-encoded' body data, while clients expected to pass the data in the body
as a JSON object. The endpoints seemed to work anyway here in the monolithic
version, though I have no idea why, but are hereby fixed properly.